Fix worldgen, update to TS, basically just a complete reqwrite making everything better because I can
