+++ .gitignore
@@ -1,0 +1,3 @@
+node_modules/
+client2/**/*.js
+client2/**/*.js.map
