+++ client/js/workerInterface.js
@@ -1,0 +1,3 @@
+$(function () {
+    
+});
