+++ client2/css/page.css
@@ -1,0 +1,89 @@
+@import url(http://fonts.googleapis.com/css?family=Inconsolata);
+
+#in, #out, #resources {
+    z-index: 1;
+}
+
+html {
+    background-color: #000;
+}
+
+*
+{
+    font-family: "Inconsolata";
+    font-size: 13px;
+}
+
+body {
+    font-family: "Inconsolata";
+    font-size: 13px;
+    color: #0f0;
+}
+
+#in {
+    width: 100%;
+    padding: 8px 0px 8px 1.25em;
+    border-color: #fff;
+    border-width: 1px 0 0 0;
+    background-color: #000;
+    color: #0f0;
+}
+
+#in:focus-visible {
+	border-width: 1px 0 0 0;
+}
+
+#bottom {
+	display: block;
+    position: fixed;
+	border-color: #fff;
+    border-width: 1px 0 0 0;
+    left: 0;
+    bottom: 0;
+    width: 100%;
+}
+#bottom::before {
+	content: "\> ";
+	margin-bottom: -1.75em;
+    display: flex;
+    z-index: 100;
+    position: relative;
+    width: 1em;
+}
+
+#out
+{
+    position: absolute;
+    padding-bottom: 13px;
+}
+
+/* World Viewer */
+#worldViewer {
+	text-align: end;
+	display:grid;
+	row-gap: 5px;
+}
+#worldViewer span {
+	padding: 10px;
+	display:  inline-grid;
+}
+.tileType01 {
+	background-color: #000;
+	color:#fff;
+}
+.tileType02 {
+	background-color: #0055ba;
+	color:#fff;
+}
+.tileType03 {
+	background-color: #60bf8a;
+	color:#000;
+}
+.tileType04 {
+	background-color: #ffc480;
+	color:#000;
+}
+.tileType05 {
+	background-color: #cf000f;
+	color:#fff;
+}
