--- js/workerInterface.js
@@ -1,3 +1,0 @@
-$(function () {
-    
-});
